VIB-638: Fix Warnings (#222)
VIB-638: Resolved warnings

VIB-638: Resolved pipeline warnings; Removed deprecated Init/Reinit

VIB-638: Participant hash to uint64_t

VIB-638: Use static_cast

VIB-638: size_t in loops; fmt pragma ignore

VIB-638: more size_t in loops; fmt pragma ignore

VIB-638: Cmake option to treat warning as errors; globally enabled warnings;

VIB-638: Fix uninitialized participantController in ITest

VIB-638: Clean CMake /W3 default in ThirdParty folder

+++ IntegrationBus/IntegrationTests/ServiceDiscoveryITest.cpp
@@ -34,7 +34,7 @@ protected:
 TEST_F(ServiceDiscoveryITest, discover_service_removal_on_participant_shutdown)
 {
     auto domainId = static_cast<uint32_t>(GetTestPid());
-    int numberOfServices = 5;
+    size_t numberOfServices = 5;
     std::string subscriberName = "Subscriber";
     std::string publisherName = "Publisher";
 
@@ -50,7 +50,7 @@ TEST_F(ServiceDiscoveryITest, discover_service_removal_on_participant_shutdown)
         ib::CreateSimulationParticipant(ib::cfg::MockParticipantConfiguration(), subscriberName, domainId, false);
 
     // Services
-    for (auto i = 0; i < numberOfServices; i++)
+    for (auto i = 0u; i < numberOfServices; i++)
     {
         const auto topic = "TopicName-" + std::to_string(i);
         publisher->CreateDataPublisher(topic, ib::sim::data::DataExchangeFormat{}, {}, 0);
